States data now can be submit through forms.
