+++ admission/src/App.js
@@ -6,6 +6,7 @@ import { Container } from "react-bootstrap"
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom"
 import Dashboard from "./components/Dashboard/Dashboard"
 import FEDSEForm from "./components/RegistrationForms/FEDSEForm"
+import {FieldsProvider} from "./components/States/FieldStates"
 
 function App() {
   return (
@@ -14,7 +15,12 @@ function App() {
         <Switch>
           <Route exact path="/" component={Auth} />
           <Route path="/Dashboard" component={Dashboard} />
+
+        {/* Added FieldsProvider so that states can be accessed in FEDSE Form from FieldsContext */}
+        <FieldsProvider>
           <Route path="/fedseform" component={FEDSEForm} />
+        </FieldsProvider>
+
         </Switch>
       </Router>
     </Container>
