+++ admission/src/components/RegistrationForms/Fields/DateField.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { Col, Form } from "react-bootstrap"
 import { useState, useContext } from "react"
-import { FieldsContext } from "../FEDSEForm"
+import { FieldsContext } from "../../States/FieldStates"
 
 export default function DateField({ title, controlId }) {
   // const [dob, setDob] = useState("")
