+++ admission/src/components/RegistrationForms/Fields/GenderField.js
@@ -3,7 +3,7 @@ import Form from "react-bootstrap/Form"
 import Row from "react-bootstrap/Row"
 import Col from "react-bootstrap/Col"
 import { useState, useContext } from "react"
-import { FieldsContext } from "../FEDSEForm"
+import { FieldsContext } from "../../States/FieldStates"
 
 export default function GenderField({ controlId }) {
   // const [studentGender, setStudentGender] = useState("")
