+++ admission/src/components/RegistrationForms/Fields/TextField.js
@@ -1,12 +1,10 @@
 import React from "react"
 import Form from "react-bootstrap/Form"
 import { useState, useContext } from "react"
-import { FieldsContext } from "../FEDSEForm"
+import { FieldsContext } from "../../States/FieldStates"
 import { Col } from "react-bootstrap"
 
 export default function TextField({ title, placeholder, controlId }) {
-  // const [motherName, setMotherName] = useState("")
-  // const [studentName, setStudentName] = useState("")
 
   const { motherNameState, studentNameState } = useContext(FieldsContext)
